requirements.txt
streamlit
tensorflow-cpu==2.13.0
numpy
pandas
scikit-learn
yfinance
plotly
joblib